+++ src/main/java/fr/alasdiablo/janoeo/worlds/block/JSaplingBlock.java
@@ -1,14 +1,10 @@
 package fr.alasdiablo.janoeo.worlds.block;
 
 import net.minecraft.block.AbstractBlock;
-import net.minecraft.block.BlockState;
 import net.minecraft.block.SaplingBlock;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.trees.Tree;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockReader;
 
 public class JSaplingBlock extends SaplingBlock {
     public JSaplingBlock(Tree treeIn, String registryName) {