+++ src/main/java/fr/alasdiablo/janoeo/worlds/init/ModFoods.java
@@ -1,0 +1,8 @@
+package fr.alasdiablo.janoeo.worlds.init;
+
+import net.minecraft.item.Food;
+import net.minecraft.item.Foods;
+
+public class ModFoods {
+    public static final Food CHERRY = Foods.APPLE;
+}